requirements.txt
opencv-python-headless
pytesseract
numpy
Pillow
fastapi
uvicorn[standard]
python-multipart